+++ mapsforge-map-awt/src/test/java/org/mapsforge/map/rendertheme/MapsforgeThemesTest.java
@@ -41,4 +41,16 @@ public class MapsforgeThemesTest {
         XmlRenderTheme xmlRenderTheme = MapsforgeThemes.OSMARENDER;
         Assert.assertNotNull(RenderThemeHandler.getRenderTheme(GRAPHIC_FACTORY, new DisplayModel(), xmlRenderTheme));
     }
+
+    @Test
+    public void motoriderTest() throws XmlPullParserException, IOException {
+        XmlRenderTheme xmlRenderTheme = MapsforgeThemes.MOTORIDER;
+        Assert.assertNotNull(RenderThemeHandler.getRenderTheme(GRAPHIC_FACTORY, new DisplayModel(), xmlRenderTheme));
+    }
+
+    @Test
+    public void motoriderDarkTest() throws XmlPullParserException, IOException {
+        XmlRenderTheme xmlRenderTheme = MapsforgeThemes.MOTORIDER_DARK;
+        Assert.assertNotNull(RenderThemeHandler.getRenderTheme(GRAPHIC_FACTORY, new DisplayModel(), xmlRenderTheme));
+    }
 }