+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 - Map rotation [#7](https://github.com/mapsforge/mapsforge/issues/7)
 - Fractional zoom [#75](https://github.com/mapsforge/mapsforge/issues/75)
     - `Parameters.FRACTIONAL_ZOOM`
+- Motorider map theme [#1483](https://github.com/mapsforge/mapsforge/issues/1483)
 - `mapsforge-themes` change package [#1135](https://github.com/mapsforge/mapsforge/issues/1135)
   - Rename `InternalRenderTheme` to `MapsforgeThemes`
 - Minor improvements and bug fixes